Fixing payment, adding detail history, review list

+++ app/src/main/java/com/android/sigemesapp/data/source/remote/response/DetailRoomResponse.kt
@@ -41,7 +41,7 @@ data class PricingItemRoom(
 data class DetailRoom(
 
 	@field:SerializedName("area_m2")
-	val areaM2: Int,
+	val areaM2: Double,
 
 	@field:SerializedName("available_slot")
 	val availableSlot: Int,
